TLG:ER:PE: newsletter components organized in directories
