+++ src/components/landingPage/LandingPage.tsx
@@ -1,4 +1,4 @@
-import NewsLetterForm from "../../newsLetter/NewsLetterForm";
+import NewsLetterForm from "../../newsLetter/form/NewsLetterForm";
 import FeaturesContent from "../featuresSection/FeaturesContent";
 import HeroContent from "../heroSection/HeroContent";
 import LandingCarousel from "../landingCarousel/LandingCarousel";
@@ -8,13 +8,12 @@ import PromoContent from "../promoBanner/PromoContent";
 const LandingPage = () => {
   return (
     <>
-  {/*     <PromoContent />
+      {/*     <PromoContent />
       <NavBarTop />
       <HeroContent />
       <FeaturesContent />
       <LandingCarousel /> */}
       <NewsLetterForm />
-    
     </>
   );
 };
