--- src/newsLetter/newsLetterForm.scss
@@ -1,125 +1,0 @@
-@import '../style.scss';
-
-.container{
-    z-index: 10;
-    background-color: $main-black;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem;
-    max-width: 100vi;
-    overflow-inline: hidden;
-    padding-block: 2rem;
-
-
-    h2{
-        font-weight: 0.8rem;
-        font-size: 0.7rem;
-        color: white;
-        @media only screen and (min-width:798px) {
-            font-size: 1.5rem;
-        }
-    }
-
-    form{
-        display: flex;
-        flex-direction: column;
-        padding-block: 2rem;
-       
-        gap: 0.8rem;
-        @media only screen and (min-width:798px) {
-               gap: 2rem;
-            }
-label {
-            color: white;
-            font-size: 0.6rem;
-    
-            @media only screen and (min-width:798px) {
-                font-size: 0.8rem;
-            }
-        }
-
-
-.input-cont {
-    display: flex;
-    flex-direction: row;
-    align-items: baseline;
-   
-
-    .submit-btn {
-            box-sizing:content-box;
-            color: $main-white;
-            font-size: 0.8rem;
-            width: 0.5rem;
-            align-self: center;
-        &:hover{
-            background-color: transparent;
-            color: grey;
-        }
-                   @media only screen and (min-width:798px) {
-                 width: 0.5rem;
-                   }
-
-        }
-
-    
-        input {
-         /*    width: 10rem; */
-            width: auto;
-            background-color: transparent;
-            color: white;
-            border: none;
-            border-bottom: 1px solid $main-white;
-            outline: none;
-            
-    
-            @media only screen and (min-width:798px) {
-                padding-block: 0.3rem;
-                width: 20rem;
-            }
-        }
-}
-
-      
-        .btn-cont {
-                display: flex;
-                gap: 0.3rem;
-                position: relative;
-
-                @media only screen and (min-width:798px) {
-                        gap: 2rem;
-                    }
-            }
-
-            .gender-btn{
-                width: 5rem;
-                color: $main-white;
-                background-color: $secondary-black;
-                font-size: 0.8rem;
-                &:hover{
-                    background-color:$main-white;
-                    color: $main-black;
-                    border: 1px solid $main-black;
-                }
-                @media only screen and (min-width: 798px) {
-                    width: 8rem;
-                }
-
-            }
-
-            .active {
-                @extend .gender-btn;
-                background-color: $main-white;
-                    color: $main-black;
-                    border: 1px solid $main-black;
-            }
-
-
-         
-
-          
-
-            }
-    }
-
